chore: adjust ui/ux of the new modal
