--- campaign-launcher/client/src/components/ConnectWallet/ConnectWalletModal.tsx
@@ -1,299 +1,0 @@
-import {
-  createElement,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-  type FC,
-} from 'react';
-
-import SearchIcon from '@mui/icons-material/Search';
-import {
-  Box,
-  Button,
-  CircularProgress,
-  Grid,
-  InputAdornment,
-  Stack,
-  TextField,
-  Typography,
-} from '@mui/material';
-import '@reown/appkit-ui/wui-qr-code';
-
-import ResponsiveOverlay from '@/components/ResponsiveOverlay';
-import { useReownWalletOptions } from '@/hooks/useReownWalletOptions';
-
-type Props = {
-  open: boolean;
-  onClose: () => void;
-};
-
-type AppKitQrCodeElement = HTMLElement & {
-  alt?: string;
-  imageSrc?: string;
-  size?: number;
-  theme?: 'dark' | 'light';
-  uri?: string;
-};
-
-type AppKitQrCodeProps = {
-  alt: string;
-  imageSrc?: string;
-  uri: string;
-};
-
-const AppKitQrCode: FC<AppKitQrCodeProps> = ({ alt, imageSrc, uri }) => {
-  const qrCodeRef = useRef<AppKitQrCodeElement>(null);
-
-  useEffect(() => {
-    if (!qrCodeRef.current) return;
-
-    qrCodeRef.current.alt = alt;
-    qrCodeRef.current.imageSrc = imageSrc;
-    qrCodeRef.current.size = 500;
-    qrCodeRef.current.theme = 'light';
-    qrCodeRef.current.uri = uri;
-  }, [alt, imageSrc, uri]);
-
-  return createElement('wui-qr-code', {
-    ref: qrCodeRef,
-    style: { display: 'block', height: '100%', width: '100%' },
-  });
-};
-
-const ConnectWalletModal: FC<Props> = ({ open, onClose }) => {
-  const [search, setSearch] = useState('');
-  const [showAllWallets, setShowAllWallets] = useState(false);
-  const {
-    connectingWallet,
-    connectWallet,
-    displayedWallets,
-    fetchMoreWallets,
-    hasMoreWallets,
-    isFetchingWallets,
-    isFetchingWcUri,
-    resetSearch,
-    resetWalletConnect,
-    wcUri,
-  } = useReownWalletOptions({ open, search, showAllWallets });
-
-  const handleClose = useCallback(() => {
-    resetWalletConnect();
-    setSearch('');
-    resetSearch();
-    setShowAllWallets(false);
-    onClose();
-  }, [onClose, resetSearch, resetWalletConnect]);
-
-  return (
-    <ResponsiveOverlay
-      open={open}
-      onClose={handleClose}
-      isLoading={isFetchingWcUri}
-      desktopSx={{
-        width: 640,
-        height: 560,
-        maxHeight: 'calc(100dvh - 48px)',
-        px: 4,
-        py: 4,
-      }}
-      mobileSx={{ height: '85dvh', p: 2 }}
-      closeButtonSx={{ top: { xs: 16, md: 32 }, right: { xs: 16, md: 32 } }}
-    >
-      {wcUri && connectingWallet ? (
-        <Stack
-          sx={{
-            alignItems: 'center',
-            gap: 2,
-            height: '100%',
-            justifyContent: 'center',
-            px: { xs: 2, md: 4 },
-            textAlign: 'center',
-          }}
-        >
-          <Typography variant="h6" sx={{ color: 'white' }}>
-            Scan with {connectingWallet.name}
-          </Typography>
-          <Box
-            sx={{
-              bgcolor: 'white',
-              borderRadius: 2,
-              height: 240,
-              p: 2,
-              width: 240,
-            }}
-          >
-            <AppKitQrCode
-              alt={connectingWallet.name}
-              imageSrc={connectingWallet.imageUrl}
-              uri={wcUri}
-            />
-          </Box>
-          <Typography variant="body2" sx={{ color: 'text.secondary' }}>
-            Open your wallet app and scan this QR code to continue.
-          </Typography>
-          <Button
-            variant="outlined"
-            onClick={() => {
-              resetWalletConnect();
-            }}
-          >
-            Back to wallets
-          </Button>
-        </Stack>
-      ) : (
-        <Stack sx={{ height: '100%', minHeight: 0 }}>
-          <Typography variant="h6" sx={{ color: 'white', mb: 1 }}>
-            Connect Wallet
-          </Typography>
-          <Typography
-            variant="body2"
-            sx={{ color: 'text.primary', fontWeight: 500, mb: 3, pr: 5 }}
-          >
-            Connect your wallet to create, participate in campaigns and even
-            track your performance on the leaderboard.
-          </Typography>
-
-          {showAllWallets && (
-            <TextField
-              value={search}
-              onChange={(event) => setSearch(event.target.value)}
-              placeholder="Search Wallet"
-              size="small"
-              fullWidth
-              sx={{ mb: 2 }}
-              slotProps={{
-                input: {
-                  endAdornment: isFetchingWallets ? (
-                    <InputAdornment position="end">
-                      <CircularProgress size={18} />
-                    </InputAdornment>
-                  ) : undefined,
-                  startAdornment: (
-                    <InputAdornment position="start">
-                      <SearchIcon sx={{ color: 'text.secondary' }} />
-                    </InputAdornment>
-                  ),
-                },
-              }}
-            />
-          )}
-
-          <Box sx={{ minHeight: 0, overflowY: 'auto', pr: 0.5, mb: 2 }}>
-            <Grid container spacing={2}>
-              {displayedWallets.map((wallet) => {
-                const isConnectingWallet =
-                  connectingWallet?.id === wallet.id && isFetchingWcUri;
-
-                return (
-                  <Grid size={{ xs: 6, md: 4 }} key={wallet.id}>
-                    <Button
-                      disabled={isFetchingWcUri}
-                      onClick={() => {
-                        void connectWallet(wallet);
-                      }}
-                      sx={{
-                        alignItems: 'center',
-                        border: '1px solid',
-                        borderColor: 'rgba(205, 199, 255, 0.22)',
-                        borderRadius: '8px',
-                        color: 'white',
-                        display: 'flex',
-                        flexDirection: 'column',
-                        gap: 1.5,
-                        height: 132,
-                        justifyContent: 'center',
-                        p: 2,
-                        width: '100%',
-                        '&:hover': {
-                          bgcolor: 'rgba(205, 199, 255, 0.08)',
-                          borderColor: 'rgba(205, 199, 255, 0.45)',
-                        },
-                      }}
-                    >
-                      {isConnectingWallet ? (
-                        <CircularProgress size={36} />
-                      ) : (
-                        <Box
-                          component="img"
-                          src={wallet.imageUrl}
-                          alt={wallet.name}
-                          sx={{
-                            borderRadius: 1.5,
-                            height: 48,
-                            objectFit: 'contain',
-                            width: 48,
-                          }}
-                        />
-                      )}
-                      <Typography
-                        variant="body2"
-                        sx={{
-                          color: 'white',
-                          maxWidth: '100%',
-                          overflow: 'hidden',
-                          textOverflow: 'ellipsis',
-                          whiteSpace: 'nowrap',
-                        }}
-                      >
-                        {wallet.name}
-                      </Typography>
-                    </Button>
-                  </Grid>
-                );
-              })}
-            </Grid>
-
-            {!isFetchingWallets && displayedWallets.length === 0 && (
-              <Stack sx={{ alignItems: 'center', py: 15 }}>
-                <Typography variant="body2" sx={{ color: 'text.primary' }}>
-                  No wallets found
-                </Typography>
-              </Stack>
-            )}
-          </Box>
-
-          <Stack
-            direction="row"
-            sx={{
-              alignItems: 'center',
-              gap: 1,
-              justifyContent: 'center',
-              mt: 'auto',
-            }}
-          >
-            {!showAllWallets ? (
-              <Button
-                variant="outlined"
-                startIcon={<SearchIcon />}
-                onClick={() => setShowAllWallets(true)}
-              >
-                Search Wallet
-              </Button>
-            ) : (
-              <>
-                <Button
-                  disabled={isFetchingWcUri}
-                  variant="outlined"
-                  onClick={() => setShowAllWallets(false)}
-                >
-                  Back
-                </Button>
-
-                <Button
-                  variant="outlined"
-                  disabled={isFetchingWallets || !hasMoreWallets}
-                  onClick={fetchMoreWallets}
-                >
-                  Show more
-                </Button>
-              </>
-            )}
-          </Stack>
-        </Stack>
-      )}
-    </ResponsiveOverlay>
-  );
-};
-
-export default ConnectWalletModal;
