+++ campaign-launcher/client/src/components/ConnectWallet/index.tsx
@@ -2,10 +2,11 @@ import { type FC } from 'react';
 
 import { Button, type ButtonProps } from '@mui/material';
 
+import ResponsiveOverlay from '@/components/ResponsiveOverlay';
 import { useConnectWalletModal } from '@/hooks/useConnectWalletModal';
 import { ConnectWalletIcon } from '@/icons';
 
-import ConnectWalletModal from './ConnectWalletModal';
+import ConnectWalletContent from './ConnectWalletContent';
 
 type ConnectWalletProps = {
   size?: ButtonProps['size'];
@@ -37,10 +38,21 @@ const ConnectWallet: FC<ConnectWalletProps> = ({ size = 'large' }) => {
         <ConnectWalletIcon sx={{ fill: 'none', width: 18, height: 18 }} />
         Connect Wallet
       </Button>
-      <ConnectWalletModal
+      <ResponsiveOverlay
         open={isConnectWalletOpen}
         onClose={closeConnectWallet}
-      />
+        desktopSx={{
+          width: 640,
+          height: 600,
+          maxHeight: 'calc(100dvh - 48px)',
+          px: 4,
+          py: 4,
+        }}
+        mobileSx={{ height: '85dvh', p: 2 }}
+        closeButtonSx={{ top: { xs: 16, md: 32 }, right: { xs: 16, md: 32 } }}
+      >
+        <ConnectWalletContent />
+      </ResponsiveOverlay>
     </>
   );
 };
